fix(F-15E): crash in UFC export functions
The Pilot and WSO UFC functions crashed due to calling the gsub method on a nil object. Fixed parentheses placement to avoid this.
Note that these functions never export anything meaningful because their UFC objects never contain the members they're looking for.
This might be because of an F-15E update.